refactor(project structure): Move utility scripts to 'scripts' directory
Moves `docblock-checker.js`, `generate-tree.js`, and `update-readme.js` from `src/utils/` to `scripts/`. This change improves project organization and aligns with standard practices for CLI tools. Also updates the paths in Github workflow files to reflect the new location of the scripts and updates paths inside the scripts to resolve to the README and package.json files.